+++ CHANGELOG.md
@@ -6,6 +6,14 @@ View layer for Hanami
 - [Ferdinand Niedermann] Ensure to set `:disable_escape` option only for Slim and don't let Tilt to emit a warning for other template engines.
 - [glaszig] Ensure partial rendering to respect `format` overriding
 
+## v1.2.1 - 2018-10-16
+### Added
+- [Luca Guidi] Introduced new, backward compatible, signature to render a layout for testing purposes (eg. `ApplicationLayout.new({ format: :html }, "contents").render`)
+
+### Fixed
+- [Luca Guidi] Ensure layout to be rendered when using HAML 5
+- [Luca Guidi] Ensure to raise `NoMethodError` when an unknown method is invoked by a view/template
+
 ## v1.2.0 - 2018-04-06
 
 ## v1.2.0.rc2 - 2018-04-06